cambiado casting de void * a uint por qint64 (se evita el error ce compilaci?n al compilar en 64 bit)

+++ YACReaderLibrary/server/lib/bfLogging/logmessage.cpp
@@ -49,7 +49,7 @@ QString LogMessage::toString(const QString& msgFormat, const QString& timestampF
     }
 
     QString threadId;
-    threadId.setNum((unsigned int)QThread::currentThreadId());
+    threadId.setNum((qint64)QThread::currentThreadId());
     decorated.replace("{thread}",threadId);
 
     // Fill in variables
